Gfx: refactor API a bit to support readbacks better
- Remove unused return of runRenderPass
- Add a method NodeRenderer::inputAboutToFinish (provisional name) which is called
  when all nodes have finished rendering on an input, but before the pass ends, in order
  to allow the node being rendered into to schedule a readback.
- Various small refactors, in particular fix a spurious release / init cycle on beginning
  because the size was not initially set.
