+++ src/plugins/score-plugin-gfx/Gfx/Graph/TextNode.cpp
@@ -172,14 +172,13 @@ private:
     }
   }
 
-  QRhiResourceUpdateBatch* runRenderPass(
+  void runRenderPass(
       RenderList& renderer,
       QRhiCommandBuffer& cb,
       Edge& edge) override
   {
     auto& mesh = TexturedQuad::instance();
     defaultRenderPass(renderer, mesh, cb, edge);
-    return nullptr;
   }
 
   void release(RenderList& r) override
